Added text under image.
Added text under the map and news image letting the user know they can click on the image to enlarge it. The word image links to the image on a new tab.

Also added a minor change to the responsive design by setting the containerResize max-width breakpoint to 380px instead of 370px.